超级大换血（未经测试）
