--- PHP/OpenTieba_Module/Opentieba.Core.BDUSS.php
@@ -1,11 +1,0 @@
-<?php
-class BDUSS // BDUSS 类
-{
-    // 王庭茂，你把这个类所有有必要的核心功能完成了Merge进master去。
-	// 大体计划如下：
-	// 实现 登录
-	// 实现 注销（这个你可以不去实现）
-	// 实现 BDUSS获取UID用户名
-	// 可自行变通！
-}
-?>
